+++ src/App.tsx
@@ -1,14 +1,14 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import LoginPage from './pages/authPage/LoginPage';
-import ForgotPassword from './pages/authPage/ForgotPassword';
-import ResetPassword from './pages/authPage/ResetPassword';
+import Login from './pages/auth/login/Login';
+import ForgotPassword from './pages/auth/forgotpassword/ForgotPassword';
+import ResetPassword from './pages/auth/resetpassword/ResetPassword';
 
 const App = () => {
   return (
     <div>
       <BrowserRouter>
         <Routes>
-          <Route path='/' element={<LoginPage />} />
+          <Route path='/' element={<Login />} />
           <Route path='/forgot-password' element={<ForgotPassword />} />
           <Route path='/reset-password' element={<ResetPassword />} />
         </Routes>
