+++ src/components/sideNav/index.tsx
@@ -1,0 +1,6 @@
+interface NavigationProps {
+  logo: string;
+}
+export const SideNavigation = ({ logo }: NavigationProps) => {
+  return <div>{logo}</div>;
+};
